feat: 예약 관리 서비스 추가, TODO: 외부요청시 EX구현
