+++ src/main/java/com/dentallink/domain/reservation/controller/ReservationController.java
@@ -1,0 +1,4 @@
+package com.dentallink.domain.reservation.controller;
+
+public class ReservationController {
+}
